move files and cleanup
