+++ src/OpenRiaServices.Hosting.Wcf/Framework/OpenRiaServices.Hosting.Wcf.csproj
@@ -1,68 +1,68 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <TargetFramework>net472</TargetFramework>
-        <DefineConstants>SERVERFX;$(DefineConstants)</DefineConstants>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <RootNamespace>OpenRiaServices.Hosting</RootNamespace>
-    </PropertyGroup>
-    <ItemGroup>
-        <Reference Include="System.ComponentModel.DataAnnotations" />
-        <Reference Include="System.configuration" />
-        <Reference Include="System.IdentityModel" />
-        <Reference Include="System.ServiceModel" />
-        <Reference Include="System.ServiceModel.Activation" />
-        <Reference Include="System.ServiceModel.Web" />
-        <Reference Include="System.Web" />
-        <Reference Include="System.Web.Extensions" />
-        <Reference Include="mscorlib" />
-        <Reference Include="system" />
-    </ItemGroup>
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
-    </ItemGroup>
-    <ItemGroup>
-        <Compile Include="..\..\OpenRiaServices.Client.Web\Framework\Data\Behaviors\MessageUtility.cs" Link="WCF\Behaviors\MessageUtility.cs" />
-        <Compile Include="..\..\OpenRiaServices.Client.Web\Framework\Data\DomainServiceFault.cs" Link="DomainServiceFault.cs" />
-        <Compile Include="..\..\OpenRiaServices.Client\Framework\QueryResult.cs" Link="QueryResult.cs" />
-        <Compile Include="..\..\OpenRiaServices.Client.Web\Framework\Data\ServiceQueryPart.cs" Link="ServiceQueryPart.cs" />
-        <Compile Update="Services\Resource.Designer.cs">
-            <AutoGen>True</AutoGen>
-            <DesignTime>True</DesignTime>
-            <DependentUpon>Resource.resx</DependentUpon>
-        </Compile>
-        <Compile Remove="GlobalSuppressions.cs" Condition="'$(RunCodeAnalysis)' != 'true' " />
-        <Compile Update="Linq\DynamicQueryable.cs">
-            <!-- This file is kept in sync with same file from another depot. Avoid changing the style. -->
-            <ExcludeFromStyleCop>true</ExcludeFromStyleCop>
-        </Compile>
-        <Compile Update="Linq\Resource.Designer.cs">
-            <AutoGen>True</AutoGen>
-            <DesignTime>True</DesignTime>
-            <DependentUpon>Resource.resx</DependentUpon>
-        </Compile>
-        <Compile Update="WCF\Resource.Designer.cs">
-            <DesignTime>True</DesignTime>
-            <AutoGen>True</AutoGen>
-            <DependentUpon>Resource.resx</DependentUpon>
-        </Compile>
-    </ItemGroup>
-    <ItemGroup>
-        <EmbeddedResource Update="Services\Resource.resx">
-            <Generator>ResXFileCodeGenerator</Generator>
-            <LastGenOutput>Resource.Designer.cs</LastGenOutput>
-            <CustomToolNamespace>OpenRiaServices.Hosting</CustomToolNamespace>
-        </EmbeddedResource>
-        <EmbeddedResource Update="Linq\Resource.resx">
-            <CustomToolNamespace>System.Linq.Dynamic</CustomToolNamespace>
-            <Generator>ResXFileCodeGenerator</Generator>
-            <LastGenOutput>Resource.Designer.cs</LastGenOutput>
-        </EmbeddedResource>
-        <EmbeddedResource Update="WCF\Resource.resx">
-            <Generator>ResXFileCodeGenerator</Generator>
-            <LastGenOutput>Resource.Designer.cs</LastGenOutput>
-        </EmbeddedResource>
-    </ItemGroup>
-    <ItemGroup>
-        <ProjectReference Include="..\..\OpenRiaServices.Server\Framework\OpenRiaServices.Server.csproj" />
-    </ItemGroup>
+  <PropertyGroup>
+    <TargetFramework>net472</TargetFramework>
+    <DefineConstants>SERVERFX;$(DefineConstants)</DefineConstants>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <RootNamespace>OpenRiaServices.Hosting</RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System.ComponentModel.DataAnnotations" />
+    <Reference Include="System.configuration" />
+    <Reference Include="System.IdentityModel" />
+    <Reference Include="System.ServiceModel" />
+    <Reference Include="System.ServiceModel.Activation" />
+    <Reference Include="System.ServiceModel.Web" />
+    <Reference Include="System.Web" />
+    <Reference Include="System.Web.Extensions" />
+    <Reference Include="mscorlib" />
+    <Reference Include="system" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="..\..\OpenRiaServices.Client.Web\Framework\Data\Behaviors\MessageUtility.cs" Link="WCF\Behaviors\MessageUtility.cs" />
+    <Compile Include="..\..\OpenRiaServices.Client.Web\Framework\Data\DomainServiceFault.cs" Link="DomainServiceFault.cs" />
+    <Compile Include="..\..\OpenRiaServices.Client\Framework\QueryResult.cs" Link="QueryResult.cs" />
+    <Compile Include="..\..\OpenRiaServices.Client.Web\Framework\Data\ServiceQueryPart.cs" Link="ServiceQueryPart.cs" />
+    <Compile Update="Services\Resource.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Resource.resx</DependentUpon>
+    </Compile>
+    <Compile Remove="GlobalSuppressions.cs" Condition="'$(RunCodeAnalysis)' != 'true' " />
+    <Compile Update="Linq\DynamicQueryable.cs">
+      <!-- This file is kept in sync with same file from another depot. Avoid changing the style. -->
+      <ExcludeFromStyleCop>true</ExcludeFromStyleCop>
+    </Compile>
+    <Compile Update="Linq\Resource.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Resource.resx</DependentUpon>
+    </Compile>
+    <Compile Update="WCF\Resource.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Resource.resx</DependentUpon>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Update="Services\Resource.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resource.Designer.cs</LastGenOutput>
+      <CustomToolNamespace>OpenRiaServices.Hosting</CustomToolNamespace>
+    </EmbeddedResource>
+    <EmbeddedResource Update="Linq\Resource.resx">
+      <CustomToolNamespace>System.Linq.Dynamic</CustomToolNamespace>
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resource.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Update="WCF\Resource.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resource.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\OpenRiaServices.Server\Framework\OpenRiaServices.Server.csproj" />
+  </ItemGroup>
 </Project>
